fix(security): upgrade Next 14.2.35 -> 16.3.0, closing the last 5 blockers
WP-72 Story 3. The dependency security gate is now GREEN:
audit-ci reports "Passed npm security audit", down from 15 blockers.

Correction to the WP-72 triage: it claimed the minimum step was 15.5.23 and
that npm's `fixAvailable: next@16.3.0` was an over-resolution artifact. That
was wrong. Measured:

                   next@14.2.35   next@15.5.23   next@16.3.0
  pins postcss     8.4.31         8.4.31         8.5.23   (>= 8.5.18, fixed)
  optional sharp   none           ^0.34.3        ^0.35.3

15.5.23 pins the same vulnerable postcss as 14 AND adds its own sharp
^0.34.3, which would have reintroduced GHSA-f88m-g3jw-g9cj - the one
internet-reachable CVE Story 1 just closed. Only 16.3.0 clears both.
16.3.0 accepts react ^18.2.0, so no React 19 migration is required.

Changes:

- next 14.2.35 -> 16.3.0. eslint-config-next held at 15.5.23 because v16
  requires eslint >=9 and a flat-config migration, which is unrelated to
  security and would balloon this change.
- next.config.mjs: experimental.serverComponentsExternalPackages moved to
  top-level serverExternalPackages. Next 16 rejects the old key.
- Pinned `--webpack` on build and dev. Next 16 defaults to Turbopack, and
  next.config.mjs documents a Tailwind purge regression under Turbopack
  (540KB vs 93KB). Preserving webpack keeps CSS output behaviour known.
- Async params migration (required from Next 15 on):
  - app/challenges/[slug]/page.tsx and opengraph-image.tsx now await params.
  - app/(share)/weekly/[week]/page.tsx is a client component and React 18.3.1
    has no React.use() to unwrap the params Promise (verified at runtime), so
    it is split into a thin async server wrapper plus share-weekly-client.tsx.
    Its test now renders the client component directly.
- tsconfig.json jsx preserve -> react-jsx, written by `next build` itself.
  Included so the file does not flip-flop on every build.

Prerender regression caught and fixed: awaiting `searchParams` in
ChallengePage silently dropped /challenges/[slug] from prerendered to
server-rendered-on-demand. The server-side searchParams read is removed,
restoring all 3 slugs to prerender-manifest.json. This is behaviour
preserving, not a shortcut - the route is prerendered (dynamicParams=false
+ generateStaticParams), so searchParams was empty at build and the server
language was always English, and ChallengeLandingContent already reads
`?lang` on the client.

Validation: lint exit 0; tsc --noEmit exit 0; full vitest suite 1416/1420
with the same 2 pre-existing lib/activation-loop.test.ts failures and no new
ones; production build exit 0; all 3 challenge slugs confirmed prerendered.

Known follow-ups, not addressed here: middleware.ts is deprecated in favour
of proxy.ts (warning only, still functional), and .audit-ci.jsonc now holds
~30 entries for advisories that no longer resolve.

Refs: Agentic OS executive-os/reviews/2026-08-07-risk-review-runsmart-web-dependency-gate.md
