+++ v0/app/challenges/[slug]/page.tsx
@@ -4,13 +4,18 @@ import { ChallengeLandingContent } from '@/components/challenge-landing-content'
 import { getAllChallengeSlugs, getChallengeTemplateBySlug, getLocalizedChallenge } from '@/lib/challengeTemplates';
 
 interface ChallengePageProps {
-  params: {
+  // Next 15+ passes this as a Promise.
+  params: Promise<{
     slug: string;
-  };
-  searchParams?: {
-    lang?: string;
-  };
+  }>;
 }
+
+// `searchParams` is deliberately not read here. Awaiting it opts the route into
+// request-time rendering, which would drop these pages out of the prerender set.
+// It was already a no-op: the route is prerendered (dynamicParams = false +
+// generateStaticParams), so searchParams was empty at build time and the server
+// language was always English. ChallengeLandingContent reads `?lang` on the
+// client and applies it there, which is what actually drove the Hebrew variant.
 
 export const dynamicParams = false;
 
@@ -18,10 +23,12 @@ export function generateStaticParams() {
   return getAllChallengeSlugs().map((slug) => ({ slug }));
 }
 
-export function generateMetadata({ params, searchParams }: ChallengePageProps): Metadata {
-  const template = getChallengeTemplateBySlug(params.slug);
-  const lang = searchParams?.lang === 'he' ? 'he' : 'en';
-  const isHebrew = lang === 'he';
+export async function generateMetadata({ params }: ChallengePageProps): Promise<Metadata> {
+  const { slug } = await params;
+  const template = getChallengeTemplateBySlug(slug);
+  // Always English server-side: these pages are prerendered, so metadata cannot
+  // vary by `?lang`. This matches what the Next 14 prerendered output already emitted.
+  const isHebrew: boolean = false;
 
   if (!template) {
     return {
@@ -32,7 +39,7 @@ export function generateMetadata({ params, searchParams }: ChallengePageProps): 
 
   // Get localized template if Hebrew
   const displayTemplate = isHebrew
-    ? { ...template, ...getLocalizedChallenge(params.slug, 'he') }
+    ? { ...template, ...getLocalizedChallenge(slug, 'he') }
     : template;
 
   const description = `${displayTemplate.tagline}. ${displayTemplate.promise}`;
@@ -43,9 +50,9 @@ export function generateMetadata({ params, searchParams }: ChallengePageProps): 
     description,
     alternates: {
       languages: {
-        'en': `/challenges/${params.slug}?lang=en`,
-        'he': `/challenges/${params.slug}?lang=he`,
-        'x-default': `/challenges/${params.slug}`,
+        'en': `/challenges/${slug}?lang=en`,
+        'he': `/challenges/${slug}?lang=he`,
+        'x-default': `/challenges/${slug}`,
       },
     },
     openGraph: {
@@ -71,9 +78,10 @@ export function generateMetadata({ params, searchParams }: ChallengePageProps): 
   };
 }
 
-export default function ChallengePage({ params, searchParams }: ChallengePageProps) {
-  const template = getChallengeTemplateBySlug(params.slug);
-  const initialLanguage = (searchParams?.lang === 'he' ? 'he' : 'en') as 'en' | 'he';
+export default async function ChallengePage({ params }: ChallengePageProps) {
+  const { slug } = await params;
+  const template = getChallengeTemplateBySlug(slug);
+  const initialLanguage: 'en' | 'he' = 'en';
 
   if (!template) {
     notFound();
