+++ v0/package.json
@@ -3,9 +3,9 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev",
+    "dev": "next dev --webpack",
     "init": "npm install",
-    "build": "next build",
+    "build": "next build --webpack",
     "start": "next start",
     "lint": "node scripts/run-eslint.cjs",
     "lint:fix": "node scripts/run-eslint.cjs --fix",
@@ -89,7 +89,7 @@
     "isomorphic-dompurify": "^2.35.0",
     "lucide-react": "^0.454.0",
     "maplibre-gl": "^5.13.0",
-    "next": "14.2.35",
+    "next": "16.3.0",
     "next-themes": "^0.4.4",
     "node-fetch": "^3.3.2",
     "posthog-js": "^1.257.0",
@@ -122,7 +122,7 @@
     "bundlesize": "^0.18.2",
     "cssnano": "^7.1.2",
     "eslint": "^8.57.1",
-    "eslint-config-next": "14.2.35",
+    "eslint-config-next": "15.5.23",
     "eslint-plugin-security": "^3.0.1",
     "fake-indexeddb": "^6.0.1",
     "jsdom": "^26.1.0",
